Fix focus mode and move functions from api.js to preload.js
